Centralize queue impls
