+++ synapse/lib/queue.py
@@ -1,9 +1,12 @@
+import asyncio
 import threading
 import collections
 
 import synapse.exc as s_exc
+import synapse.glob as s_glob
 import synapse.eventbus as s_eventbus
 
+import synapse.lib.base as s_base
 
 class Queue(s_eventbus.EventBus):
     '''
@@ -240,3 +243,176 @@ class Queue(s_eventbus.EventBus):
                 yield self.slice(size, timeout=timeout)
             except s_exc.IsFini as e:
                 return
+
+class AsyncQueue(s_base.Base):
+    '''
+    Multi-async producer, single sync finite consumer queue with blocking at empty and full.
+    '''
+    async def __anit__(self, max_entries, drain_level=None):
+        '''
+        Args:
+            max_entries (int): the maximum number of entries that can be in the queue.
+
+            drain_level (int): once the queue is full, no more entries will be admitted until the number of entries
+                falls below this parameter.
+        '''
+
+        await s_base.Base.__anit__(self)
+        self.deq = collections.deque()
+        self.notdrainingevent = asyncio.Event(loop=self.loop)
+        self.notdrainingevent.set()
+        self.notemptyevent = threading.Event()
+        self.max_entries = max_entries
+        self.drain_level = max_entries // 2 if drain_level is None else drain_level
+
+        async def _onfini():
+            self.notemptyevent.set()
+            self.notdrainingevent.set()
+        self.onfini(_onfini)
+
+    def get(self):
+        '''
+        Pending retrieve on the queue
+        '''
+        while not self.isfini:
+            try:
+                val = self.deq.popleft()
+                break
+            except IndexError:
+                self.notemptyevent.clear()
+                if len(self.deq):
+                    continue
+                self.notemptyevent.wait()
+        else:
+            return None
+
+        if not self.notdrainingevent.is_set():
+            if len(self.deq) < self.drain_level:
+                s_glob.plex.callSoonSafe(self.notdrainingevent.set)
+        return val
+
+    async def put(self, item):
+        '''
+        Put onto the queue.  It will async pend if the queue is full or draining.
+        '''
+        while not self.isfini:
+            if len(self.deq) >= self.max_entries:
+                self.notdrainingevent.clear()
+            if not self.notdrainingevent.is_set():
+                await self.notdrainingevent.wait()
+                continue
+            break
+        else:
+            return
+
+        self.deq.append(item)
+        self.notemptyevent.set()
+
+class AQueue(s_base.Base):
+    '''
+    An async queue with chunk optimized sync compatible consumer.
+    '''
+    async def __anit__(self):
+        await s_base.Base.__anit__(self)
+
+        self.fifo = []
+        self.event = asyncio.Event(loop=self.loop)
+        self.onfini(self.event.set)
+
+    def put(self, item):
+        '''
+        Add an item to the queue.
+        '''
+        if self.isfini:
+            return False
+
+        self.fifo.append(item)
+
+        if len(self.fifo) == 1:
+            self.event.set()
+
+        return True
+
+    async def slice(self):
+
+        # sync interface to the async queue
+        if len(self.fifo) == 0:
+            await self.event.wait()
+
+        retn = list(self.fifo)
+        self.fifo.clear()
+        self.event.clear()
+        return retn
+
+class S2AQueue(s_base.Base):
+    '''
+    Sync single producer, async single consumer finite queue with blocking at empty and full.
+    '''
+    async def __anit__(self, max_entries, drain_level=None):
+        '''
+        Args:
+            max_entries (int): the maximum number of entries that can be in the queue.
+
+            drain_level (Optional[int]): once the queue is full, no more entries will be admitted until the number of
+                entries falls below this parameter.  Defaults to half of max_entries.
+        '''
+
+        await s_base.Base.__anit__(self)
+        self.deq = collections.deque()
+        self.notdrainingevent = threading.Event()
+        self.notdrainingevent.set()
+        self.notemptyevent = asyncio.Event(loop=self.loop)
+        self.max_entries = max_entries
+        self.drain_level = max_entries // 2 if drain_level is None else drain_level
+        assert self.drain_level
+
+        async def _onfini():
+            self.notemptyevent.set()
+            self.notdrainingevent.set()
+        self.onfini(_onfini)
+
+    async def get(self):
+        '''
+        Async pend retrieve on the queue
+        '''
+        while not self.isfini:
+            try:
+                val = self.deq.popleft()
+                break
+            except IndexError:
+                self.notemptyevent.clear()
+                if len(self.deq):
+                    continue
+                await self.notemptyevent.wait()
+        else:
+            return None  # Raise fini?
+
+        if not self.notdrainingevent.is_set() and len(self.deq) < self.drain_level:
+            self.notdrainingevent.set()
+
+        return val
+
+    def put(self, item):
+        '''
+        Put onto the queue.  Pend if the queue is full or draining.
+        '''
+        while not self.isfini:
+            if len(self.deq) >= self.max_entries:
+                self.notdrainingevent.clear()
+            if not self.notdrainingevent.is_set():
+                self.notdrainingevent.wait()
+                continue
+            break
+        else:
+            return
+
+        self.deq.append(item)
+
+        # N.B. asyncio.Event.is_set is trivially threadsafe, though Event.set is not
+
+        if not self.notemptyevent.is_set():
+            self.loop.call_soon_threadsafe(self.notemptyevent.set)
+
+    def __len__(self):
+        return len(self.deq)
+
