+++ synapse/tests/test_axon.py
@@ -1,5 +1,4 @@
 import os
-import time
 import hashlib
 import asyncio
 import logging
@@ -8,8 +7,6 @@ import synapse.exc as s_exc
 import synapse.axon as s_axon
 import synapse.glob as s_glob
 import synapse.common as s_common
-
-import synapse.lib.threads as s_threads
 
 from synapse.tests.utils import alist
 import synapse.tests.utils as s_t_utils
@@ -325,36 +322,3 @@ class AxonTest(s_t_utils.SynTest):
     async def test_axon_blobstors_dropping(self):
         # FIXME
         pass
-
-class _AsyncQueueTest(s_t_utils.SynTest):
-    async def test_asyncqueue(self):
-
-        # The axon tests test most of the asyncqueue functionality.  We just need to test the
-        # draining part
-
-        async with await s_axon._AsyncQueue.anit(5, drain_level=3) as q:
-            [await q.put(i) for i in range(5)]
-            got_to_end = False
-            last_msg = 0
-
-            def sync_worker():
-                nonlocal got_to_end
-                nonlocal last_msg
-                time.sleep(0.1)
-
-                last_msg = q.get()
-                last_msg = q.get()
-                time.sleep(0.1)
-                last_msg = q.get()
-
-                got_to_end = True
-            t = s_threads.worker(sync_worker)
-            before = time.time()
-            await q.put(6)
-            self.lt(0.1, time.time() - before)
-            self.eq(last_msg, 2)
-            await s_glob.plex.sleep(0.1)
-
-            self.true(got_to_end)
-
-            t.join()
